+++ src/components/Card/Card.tsx
@@ -57,6 +57,8 @@ export interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
   onClick?: () => void;
   /** aria-label */
   ariaLabel?: string;
+  /** Button type. Default er 'button' */
+  buttonType?: string;
 }
 
 export interface CardState {
@@ -80,7 +82,8 @@ export default class Card extends React.PureComponent<CardProps, CardState> {
     actions: null,
     marginbottom: '2px',
     margin: 'medium',
-    circleOnIcon: true
+    circleOnIcon: true,
+    buttonType: 'button'
   };
 
   constructor(props: CardProps) {
@@ -106,6 +109,7 @@ export default class Card extends React.PureComponent<CardProps, CardState> {
       className,
       circleOnIcon,
       id,
+      buttonType,
       ...props
     } = this.props;
 
@@ -144,6 +148,7 @@ export default class Card extends React.PureComponent<CardProps, CardState> {
                   ariaLabel={title}
                   circle={circleOnIcon}
                   onClick={this._toggleExpand}
+                  type={buttonType}
                 />
               </div>
             )}