+++ src/components/RadioButtonGroup/RadioButtonGroup.tsx
@@ -1,6 +1,7 @@
 import classnames from 'classnames';
 import {
   ChoiceGroup as FabricChoiceGroup,
+  IChoiceGroupOption,
   IChoiceGroupProps
 } from 'office-ui-fabric-react/lib-commonjs/ChoiceGroup';
 import * as React from 'react';
@@ -8,6 +9,10 @@ import ErrorMessage from '../ErrorMessage/ErrorMessage';
 import { getClassNames } from './RadioButtonGroup.classNames';
 import LabelWithCallout, { calloutState } from '../LabelWithCallout';
 import { LabelWithCalloutProps } from '../LabelWithCallout/LabelWithCallout';
+
+export interface IRadioButtonGroupOptions extends IChoiceGroupOption {
+  description?: string;
+}
 
 export interface RadioButtonGroupProps extends IChoiceGroupProps {
   calloutFloating?: boolean;
@@ -25,6 +30,7 @@ export interface RadioButtonGroupProps extends IChoiceGroupProps {
     oldCalloutState: calloutState,
     newCalloutState: calloutState
   ) => void;
+  options: IRadioButtonGroupOptions[];
 }
 
 /**
@@ -43,29 +49,63 @@ const RadioButtonGroup = (props: RadioButtonGroupProps) => {
     label,
     labelCallout,
     onCalloutToggle,
+    options,
     ...rest
   } = props;
+  let tempOptions = options;
+
+  if (options) {
+    options.forEach(option => {
+      if (option.description) {
+        option.onRenderLabel = DescriptionRender(option.description);
+      }
+    });
+    tempOptions = options;
+  }
+
+  const styles = getClassNames({ ...props });
 
   return (
-    <>
+    <fieldset id={id} className={styles.fieldset}>
       <LabelWithCallout
-        id={id}
         label={label}
         help={help}
         warning={warning}
+        inFieldset={true}
         calloutFloating={calloutFloating}
         onCalloutToggle={onCalloutToggle}
         {...labelCallout}
       />
       <FabricChoiceGroup
+        options={tempOptions}
         {...rest}
-        className={classnames(getClassNames(props), className)}
+        className={classnames(styles.radioButtons, className)}
         ariaLabelledBy={label}
       >
         {children}
       </FabricChoiceGroup>
       {errorMessage && <ErrorMessage>{errorMessage}</ErrorMessage>}
-    </>
+    </fieldset>
+  );
+};
+
+const DescriptionRender = (description: string) => (p: any) => {
+  return (
+    <div
+      style={{
+        display: 'flex',
+        flexDirection: 'column'
+      }}
+    >
+      <span id={p.labelId} className="ms-ChoiceFieldLabel">
+        {' '}
+        {p.text}{' '}
+      </span>
+      <span className={'descriptionLabel ms-ChoiceFieldLabel'}>
+        {' '}
+        {description}{' '}
+      </span>
+    </div>
   );
 };